+++ PROYECTO-DBD/app/Http/Controllers/CuentaBancariaController.php
@@ -3,7 +3,6 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-#use App\Models\Cuenta_bancaria;
 use App\Models\CuentaBancaria;
 use App\Models\Cliente;
 use App\Models\Feriante;
@@ -19,7 +18,6 @@ class CuentaBancariaController extends Controller
     public function index()
     {
         //
-        #$cuenta_bancarias = Cuenta_bancaria::all();
         $cuenta_bancarias = CuentaBancaria::all();
         if($cuenta_bancarias != NULL){
             return response()->json($cuenta_bancarias);
@@ -36,15 +34,15 @@ class CuentaBancariaController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
+    //Corexion
     public function store(Request $request)
     {
                 //
-        #$cuenta_bancarias = new Cuenta_bancaria();
         $cuenta_bancarias = new CuentaBancaria();
         $validatedData = $request->validate([
             'numero_cuenta' => ['required' ,'numeric'],
             'banco' => ['required' , 'min:2' , 'max:50'],
-            'tipo_cuenta' => ['required' , 'min:2' , 'max:50']
+            'tipo_cuenta' => ['required' , 'min:2' , 'max:50'],
 			
 			'id_feriante' => ['required' , 'numeric'],
             'id_cliente' => ['required' , 'numeric'],
@@ -57,20 +55,20 @@ class CuentaBancariaController extends Controller
         if($cliente == NULL){
             return response()->json([
                 'message'=>'No existe un cliente con esa id'
+            ]);
         }
         $feriante = Feriante::find($request->id_feriante);
         if($feriante == NULL){
             return response()->json([
                 'message'=>'No existe un feriante con esa id'
+            ]);
         }
-        #$orden_de_pago = Orden_de_pago::find($request->id_orden_pago);
         $orden_de_pago = OrdenDePago::find($request->id_orden_pago);
         if($orden_de_pago == NULL){
             return response()->json([
                 'message'=>'No existe orden de pago con esa id'
+            ]);
         }
-		
-		
 		
         $cuenta_bancarias->numero_cuenta = $request->numero_cuenta;
         $cuenta_bancarias->banco = $request->banco;
@@ -90,9 +88,20 @@ class CuentaBancariaController extends Controller
      */
     public function show($id)
     {
-        #$cuenta_bancarias = Cuenta_bancaria::find($id);
-        $cuenta_bancarias = CuentaBancaria::find($id);
-        return response()->($cuenta_bancarias);
+        if(is_numeric($id)){
+            $cuenta_bancarias = CuentaBancaria::find($id);
+            if($cuenta_bancarias == NULL){
+                return response()->json([
+                    'message'=>'No se encontro la cuenta bancaria'
+                ]);
+            }
+            return response()->json($cuenta_bancarias);
+        }
+        else{
+            return response()->json([
+                'message'=>'id invalido'
+            ],404);
+        }
     }
 
     /**
@@ -104,19 +113,32 @@ class CuentaBancariaController extends Controller
      */
     public function update(Request $request, $id)
     {
-        #$cuenta_bancarias = Cuenta_bancaria::find($id);
-        $cuenta_bancarias = CuentaBancaria::find($id);
-        if($request->numero_cuenta != NULL){
-            $cuenta_bancarias->numero_cuenta = $request->numero_cuenta;
+        if(is_numeric($id)){
+            $cuenta_bancarias = CuentaBancaria::find($id);
+            if($cuenta_bancarias == NULL){
+                return response()->json([
+                    'message'=>'No se encontro la cuenta bancaria'
+                ]);
+            }
+            else{
+                if($request->numero_cuenta != NULL){
+                    $cuenta_bancarias->numero_cuenta = $request->numero_cuenta;
+                }
+                if($request->banco != NULL){
+                    $cuenta_bancarias->banco = $request->banco;
+                }
+                if($request->tipo_cuenta != NULL){
+                    $cuenta_bancarias->tipo_cuenta = $request->tipo_cuenta;
+                }
+                $cuenta_bancarias->save();
+                return response()->json($cuenta_bancarias);
+            }  
         }
-        if($request->banco != NULL){
-            $cuenta_bancarias->banco = $request->banco;
+        else{
+            return response()->json([
+                'message'=>'id invalido'
+            ],404);
         }
-        if($request->tipo_cuenta != NULL){
-            $cuenta_bancarias->tipo_cuenta = $request->tipo_cuenta;
-        }
-        $cuenta_bancarias->save();
-        return response()->json($cuenta_bancarias);
     }
 
     /**
@@ -127,10 +149,26 @@ class CuentaBancariaController extends Controller
      */
     public function destroy($id)
     {
-        $cuenta_bancarias->delete();
-        return response()->json([
-            "message"-> "cuenta bancaria eliminada",
-            "id"=>$cuenta_bancarias->id
-        ],201);    
+        if(is_numeric($id)){
+            $cuenta_bancarias = CuentaBancaria::find($id);
+            if($cuenta_bancarias != NULL){
+                $cuenta_bancarias->delete();
+            }
+            else{
+                return response()->json([
+                    "message" => "id Cuenta bancaria inexistente"
+                ],404);
+            }
+
+            return response()->json([
+                "message"=> "cuenta bancaria eliminada",
+                "id"=>$cuenta_bancarias->id
+            ],201); 
+        }
+        else{
+            return response()->json([
+                'message'=>'id invalido'
+            ],404);
+        } 
     }
 }
